Split RouterContext from RouterProvider
