+++ client/components/providers/RouterProvider.js
@@ -1,15 +1,7 @@
 import { FlowRouter } from 'meteor/kadira:flow-router';
 import { Tracker } from 'meteor/tracker';
-import React, { createContext, useContext, useEffect, useMemo, useState } from 'react';
 
-export const RouterContext = createContext({
-	navigateTo: () => {},
-	replaceWith: () => {},
-	getRouteParameter: () => {},
-	watchRouteParameter: () => {},
-	getQueryStringParameter: () => {},
-	watchQueryStringParameter: () => {},
-});
+import { RouterContext } from '../../contexts/RouterContext';
 
 const navigateTo = (pathDefinition, parameters, queryStringParameters) => {
 	FlowRouter.go(pathDefinition, parameters, queryStringParameters);
@@ -51,35 +43,5 @@ const router = {
 };
 
 export function RouterProvider({ children }) {
-	return <RouterContext.Provider value={router}>
-		{children}
-	</RouterContext.Provider>;
+	return <RouterContext.Provider children={children} value={router} />;
 }
-
-export const useRoute = (pathDefinition) => {
-	const { navigateTo, replaceWith } = useContext(RouterContext);
-
-	return useMemo(() => {
-		const navigate = (...args) => navigateTo(pathDefinition, ...args);
-		navigate.replacingState = (...args) => replaceWith(pathDefinition, ...args);
-		return navigate;
-	}, [navigateTo, replaceWith]);
-};
-
-export const useRouteParameter = (name) => {
-	const { getRouteParameter, watchRouteParameter } = useContext(RouterContext);
-	const [parameter, setParameter] = useState(getRouteParameter(name));
-
-	useEffect(() => watchRouteParameter(name, setParameter), [watchRouteParameter, name]);
-
-	return parameter;
-};
-
-export const useQueryStringParameter = (name) => {
-	const { getQueryStringParameter, watchQueryStringParameter } = useContext(RouterContext);
-	const [parameter, setParameter] = useState(getQueryStringParameter(name));
-
-	useEffect(() => watchQueryStringParameter(name, setParameter), [watchQueryStringParameter, name]);
-
-	return parameter;
-};
