+++ client/hooks/useEmbeddedLayout.js
@@ -1,3 +1,3 @@
-import { useQueryStringParameter } from '../components/providers/RouterProvider';
+import { useQueryStringParameter } from '../contexts/RouterContext';
 
 export const useEmbeddedLayout = () => useQueryStringParameter('layout') === 'embedded';
